fix melee lunge code applying too many deceleration ticks
melee lunge code cleanup

+++ xlive/Blam/Engine/Game/GameTimeGlobals.cpp
@@ -39,7 +39,7 @@ int time_globals::get_ticks_difference()
 
 float time_globals::get_ticks_difference_real()
 {
-	return (get()->ticks_per_second / 30.0f);
+	return ((float)get()->ticks_per_second / 30.0f);
 }
 
 float time_globals::get_ticks_fraction_leftover()
